IMPORTANT CHANGE: Functions now resolve through the lexical context by preference
This turns out to match intuitions better, and is crucially easier for the system to disambiguate.

This change also results in a crucial bug-fix, because Functions resolve slightly *differently* for the lexical context. If it is resolved lexically, the Function is called once, with the full received context as input. If resolved through the received context, the Function is called on *each* element of the context.

I believe this is more generally correct, but keep an eye out for bugs.