Major enhancements to Extract App, to work with Model Types
Core of the problem was that that Model Types had neither the notion of their SpaceId (which is a long-standing bug that just didn't surface until now), nor their Model (which we need for shadowing). This involved a lot of refactoring of SystemType to allow this.

Removed the fancy code for hollowing Model Properties, because I realized it was wrong: we want the local Property to point to the local Type which points to the local Model Shadow. (I think.)

Also, fixed a bug that would crash if you tried to extract again from the same Space (due to duplicate backup tables).