_instances is now usually pre-sorted
We pre-calculate the inheritance tree, and pre-sort the children at each level, with the result that _instances is sorted for free *if* the Instances have Display Name or Link Name.

If the Instances rely on Computed Name, they still need to be manually sorted. But you're an advanced user by definition if you're using Computed Name.

This has the nice side-effect of making descendants() much faster on average, since we are now pre-computed the tree -- it's O(n) to build the cache, then O(log n) to calculate the descendants each time, instead of O(n log n) each call.