Make ElemValue.get[Opt]() coercion-sensitive
This is the first step towards adding MultipleChoiceType. We need this to work, so that all of those places that get a LinkType can work with MultipleChoice.

Did a long-overdue refactoring of ElemValue while I was at it.

Had to add one specific limitation, to prevent other Text Types from coercing to ErrorTextType. That's because we use the test of whether a value is ErrorText *heavily* in the standard Invocation chain. Basically, we need to treat ErrorText as semantically different from Text.