Remove PageManager.currentPage
Realized that yesterday's changes introduced a hard concept of *the* Page, which I don't want because the Help system is going to want a second Page. So switched everything to use findPageFor() instead.

Wound up needing to do a *lot* of bulletproofing around findParentGadget(), which had a bunch of nasty failure conditions. It's now much more correct.