Enhanced _if to cope with an empty predicate
While messy from a type POV, it's just plain convenient for _if() to produce false when it gets an empty result. So it's now slightly redefined, to be a bit more loosely-typed and C-like, to produce false if the predicate is *anything other than true*.

Added unit tests for this.