On the Explore Page, Make a Page now includes the current Thing
Make a Page has apparently always been broken, in the sense that it didn't take the context Thing into account, so queries that depend on that didn't do the right thing. It now works correctly.

This isn't quite ideal yet -- it *should* be only including the context Thing if it is relevant to the query. But that's not easy to calculate, and having it present doesn't *hurt*, so let's play it safe.