Several test fixups
Removed a couple of things that were causing excess spews

Removed the quit() at the end of the functional test -- it seems to happen automatically, and saying it explicitly causes connection problems.

Added an extra wait in the created-thing situation, but I'm honestly unsure why I need it. It's worrying.