.3y28an1: we no longer over-escape entities in the page title
The fix for this was fairly easy: we are now transmitting the raw form, but only using it in this one place.

The *vast* majority of this checkin is the regression test for this bug. Added lots more machinery to the functional test harness to be able to do so, including the ability to create a Thing.