If there is an error in a multi-phrase expression, that error prevents evaluation of the rest of the expression.
This is more hard-assed than we used to be, but I think it's correct.

Also, switched ErrorType to be PlainText -- I have no idea why it was QLText, but that was dumb and dangerous: displaying the error required more QL processing, which could lead to recursive errors.

Also, fixed a bug that was displaying the type in the error. Not sure why I only just noticed this.