.3y28b5z: _QLButton/Link/Input now accept more or less any Type, not just Things, and should work as expected.
Added the _uniqueHtmlId function, to make it easier to have a lot of _QLButtons on a page, each with its own target.

Enhanced TextTypes to coerce automatically to ParsedTextType. Main upshot of this is that you can now use Properties directly in some parameters that previous wouldn't allow this. (For example, in the label parameter of _QLButton.)