.3y28aby: made it less likely that the top navbar will splay onto two lines at 900 pixels.
Basically hacked Bootstrap, which had a hardcoded switch at 992px that caused the navbar's container to clamp down to 750px, shoving everything into two lines. Now, we only go to two lines when the header truly runs out of space, making this problem a little less common, although it can still happen.

To help this, removed the "Logged in as" prefix on the right-hand menu -- most modern systems don't bother.