Progress towards updating the Space with the new Permission Defaults Thing
To make this work properly, I need to be able to wait for the evolution to finish while starting. Wound up reworking the main create and modify functions to be properly RequestM-based, rather than the side-effecting horror they had been. It's a lot of dangerous changes, but the result is much cleaner.