.3y28625, .3y28abz, .3y287el: massive improvements to the streaming-upload protocol
This protocol was *very* unreliable, and no longer worked well once we moved to the clustered environment. It has been rewritten, to add a simplistic but theoretically reliable streaming protocol between the front and back ends.

Also, fixed an idiotic race bug in PhotoUploadActor that was frequently causing the "I'm done" message to fail to get back to the front end. This means that you now usually see the upload complete properly.