Fixes to get the QuerkiNodeCoordinator serializing its snapshots properly
The minor fix was to the timing of taking the snapshots, so they actually happen.

The major fix was the realization that Sets and Maps do *not* automatically serialize -- I have to register all of those types as well! Spent quite a while figuring out how best to do this.