A lot of changes to the SpaceCore, to fix a subtle problem in Space creation.
The bug was that, when I created a new Space, the page came up with none of the creation buttons or conversations. This was because the UserSpaceSession was getting a partially-created SpaceState, that didn't yet have the Owner Person, so _hasPermission() was failing.

Wound up doing a lot of adjustment to the timing and flows of that Person-creation code, including an enhancement to Requester to let me punch RequestM through persist(). The result should be more correct.