Import Space now works correctly in the Akka Persistence world
Almost none of these changes are to import itself, which works just fine. The problem is, the old system would create the new Space, *leave it running*, and then create an alternate Actor for persisting the import process. That's highly illegal in the new Akka Persistence world. So we needed to formalize the concept of "space status", and introduce a new StatusInitializing, to at least mostly prevent the real Space from running while we are finishing the import.