.7w4g7ww: fixed the fact that the page doesn't reload properly when I log in
The fix here is kind of a punt. The issue is something weird in trying to parse the response in upickle. But in fact, as soon as we reload, the metadata system will update the user data *anyway*, so there's no good reason to even bother with this. I do wish I understood the problem, though.