.7w4g865: make debugging of typos less sucktastic
This is a small but critical semantic change to QL. Previously, you could put any name into a QL expression, and it would be valid -- if it was unknown, it would just be treated as a Tag. (Originally to be "wikiish".) This meant that typos tended to get buried, and were hard to detect.

Now, we only do this if it is either the name of an actual Thing, or a Tag that is being used somewhere. If not, this raises an error, so the typo comes immediately to light.

Also, tweaked things so that a single error doesn't trash the entire page -- it only errors out of this QL expression, and is displayed in the context of the rest of the page, which is rendered normally. This makes it easier to see where the error is.

Between them, these changes should make debugging quite a bit easier.