Checkpoint: much of the way into rewriting SpaceCore to being purer.
We have a building problem, that operations involving multiple actions aren't atomic. Worse, we have no good path towards being able to make *big* changes to a Space, and that's coming.

So we're moving towards a purer and more composable style in SpaceCore, which should allow complex actions to work naturally.