Fixed some peculiarities in emails produced by _emailLink.
The issue here turned out to be that the QText engine was just plain broken, because of the way Actuarius built in a *default* Decorator and used that in a rather ad-hoc way. Stripped that out, and rewrote a bunch of code as necessary to force Decorator to get properly passed down from the top.

Also removed the special-case pathway for Wikitext.email, which is no longer needed.