Optimization: memoize the parse results from parsing QLText.
This is basically dealing with some low-hanging fruit. Parsing QLText makes up a large fraction of the time required to render a complex page, and much of that parsing is redundant -- we've been parsing the same QLText over and over again. That's silly -- the QLText object is immutable, so it memoizes well. So we are doing so now, which considerably speeds up page rendering in some cases.