QI.9v5kemx: when you try to dereference an unknown name in QL, give an error
This replaces the old functionality, where it was an error to have an unknown name to begin with. That was inconvenient and inconsistent, but we still want it to be an error when you try to *use* a typo.

Note that it is still legit to deference a Tag *value* from a Property -- since a Prop Value may contain many Tags, some defined and some not, we don't want to casually make those errors.

Along the way, this is the last step in addressing QI.7w4g9nv.