QI.9v5keyo: fix the "dots" on the carousel, and don't show them if it's a single photo
The bug was mainly to ask for the dots to be suppressed because they are confusing with a single image, which was a reasonable request. But more importantly, there turns out to be an off-by-one bug that has been in this code from the outset, which Bootstrap (unfortunately) did something too reasonable with. That is now fixed, so that the number of dots actually matches the number of images.