QI.7w4gd71 -- introduced the _positions() function
This is an abstract signature, only implemented for Text types for now. Given a sequence of some sort, it produces the indexes of where a given subsequence is found within that sequence.

This can be easily used to test whether a substring is found in a string, in combination with _isEmpty.

To support this, ParsedTextType now implements IsTextType, helping to unify that with the real Text types under some circumstances.

Heavily unit-tested.

Also introduced a new function to make it easier to build a Collection. Should have done that one years ago.