Checkpoint: start more-sophisticated state evolution
This is nowhere *near* well-enough tested yet, mind. We need to confirm that we are actually using the intelligent evolution code, that it's reliably doing the right things, and that all of the TODO edge cases are handled. But it passes the existing tests, and at least hypothetically should be far more efficient.