Checkpoint: user-session Space evolution is starting to work
Tweaks to get this actually functioning as expected. (Most importantly, actually passing the events along.)

Introduced some test mechanisms for checking whether we are actually evolving efficiently or filtering the whole thing, but those don't work yet: the current machinery winds up doing that asynchronously, so we can't test it deterministically. Need to rewrite so that this is properly deterministic. (And more efficient.)

Upgraded to newer versions of Play, ScalaTest+Play and Akka, so that I could get ScalaTest 3.0.0 and the Position mechanism, which allows me to use higher-level test functions that report line numbers of errors correctly.