Checkpoint: evolution starting to work nicely
At this point, we are now able to unit-test whether an action requires a full filtering or works with an efficient evolution, and the few tests we're doing so far work properly. Still needs some adjustment, but it's looking pretty good.

Note the change in design here: we are back to doing all of the evolution at pull time, not at push time. This should be yet more efficient, and is much more deterministic for unit testing. (Slightly worse in terms of latency, but no worse than what we had before, and often far better.)