QI.bu6of67: Make view evolution cope correctly with Role assignments
Includes a full unit test; this one was done with TDD. (Yay for the Mid-Test Harness!)

Added Mid-test functions to create and grant Roles.

Fixed login function to cache the user's login results, and added a function to fetch the OID for a user.

Added Mid-test machinery to set a Property to a List of values, not just one.

Enhanced getPropertyValues() to support Link values.