Checkpoint: in the middle of adding the ExperimentActor and SpaceExperiments
Next step is figuring out how to refactor the relevant bits out of SpaceCore so that ExperimentActor can process SpaceMessages intelligently. This likely involves a deep rethink of Publication, merging it with Experiments, so that it isn't massively corrupting the SpaceCore code. See "Experiments and Publication" in Querki Design Notes for more thoughts on how this might work.

In the meantime, reorienting to get basic GraphQL implemented, since that is more critical-path.