Tags are now working correctly
Tags are weird, and probably violate basic principles of GraphQL, because the way are interpreted depends on how they are called. You can invoke them as simple primitives, and the simple name will be returned. Or, if you have reason to believe they are reified Things, you can invoke them with sub-selections, and it will dereference through them the same as with Links.