Added plumbing for GraphQL
We now have a controller, and unit tests, for the real entry point. I think we're ready to roll...

Made the controller stack (which is old and a bit crufty) more appropriately typeful, so that application-level controllers can be explicit about the parser and (more immediately relevant) can know in a properly-typed way what the body type is.

Note that we specifically ask for the content-type to be "application/graphql" -- we're not enforcing that in a hard-assed way, but you can't just use, eg, application/text.

GraphQL is now also exposed at the API level, although nothing is using that yet.