QI.bu6ofkm: Space Tracing and the Show Space Id command
Most importantly, this adds a new config var, `querki.debug.space.[spaceId].trace` -- setting this to true causes vast amounts of spewage for that Space.

Also, introduced the `Show Space Id(ownerHandle, spaceName)` command, so that I can find the OID of a Space *without* loading it. This is so that I can disable a problematic Space (as I'm dealing with now) without needing to load it. (This should rarely be needed, but is helpful in the current situation, where loading the Space is clearly causing a severe infinite loop.)