QI.7w4g99h: add unit tests for basic publication
This ticket is about adding QL support for publication functions. In order to prove that out, let's start by finally adding a little test support for the functions to begin with. These were completely untested before, but now that we have the mid-level harness it's reasonably straightforward to add it.

Note that, as part of this, we've extracted some useful functions from SecurityMidTests into SecurityMidFuncs, so that we can reuse them.