QI.bu6ofkm: Introduce the new /admin/dumpreadable/:spaceId entry point
This is a super-specialized page, only usable by Admins. There is no attempt to make it easy to use, because its use should be super-rare. It loads the specified Space, using the History mechanism instead of the standard Space pathways, and dumps the whole thing in QLog format.

The goal here is so that, when a Space is truly *broken*, we can visually inspect it and try to puzzle out what is going on.

This page simply bounces if you try to load it and aren't a logged-in Admin.