QI.bu6ofru: fix the mind-blowingly stupid bug that allows recursive expressions to destroy the entire cluster
The summary here is that we have always checked for this sort of recursion -- and then we never actually *used* the results of that check! So infinite recursion was possible, and causes everything to explode.

The fix is simply to do what this code was always supposed to do. (And possibly did at some point: not clear whether this was always broken.)