QI.3y28adl: Allow hooking the CheckList control to an RxInput
This is now getting to the point of actually working as intended, which is fun and cool. The code's not as elegant as I might wish, but it's a step in the right direction.

Introduces the "GadgetListeners" concept, which I don't appear to have had yet, allowing Gadgets around the page to talk to each other.