QI.bu6oc49: get shadowed Stylesheets working
This fixes two bugs.

One is that Stylesheets fall into an odd little niche: sort of like Pages, they are individual Things that should nonetheless get shadowed in child Spaces. (I suspect we'll find other such examples as we go.)

The second is that, when listing choices to pick from, we were explicitly filtering shadows out! I'm not sure why we were doing that, but it had meant that Stylesheets weren't showing up when you wanted to pick one.

Along the way, improved the way we filter candidates to display when picking from a LinkModel: this now uses the modern cached descendants() system, so it's no longer n**2.