QI.9v5kfif: now fixed
As expected, the fix was basically a one-liner (in SpaceMembersActor), but it took quite a lot of test enhancement to get to the point of being able to prove that. The test is now properly clean and clear.

This introduces the concept of "guest" more formally into the test harness, and gets the signup for that right. (Which was most of the effort.) Wound up finally dealing with fixing `getPersonInfoFor()` correctly, by adding a new `getMyInfo()` API call -- this is only being used by the tests so far, but it's a reasonable thing to have in the API.