QI.7w4ger4: Make Computed Name win over Link Name
Once upon a time, Link Name was the primary name for a Thing. That was many evolutions ago, but we wound up in a position where Link Name overrides Computed Name. Eric pointed out that that seems wrong, and I think he's right. So this rewrites the name code so that we check (in order):

* Explicit Display Name
* Computed Display Name
* Link Name
* OID

and render whichever one comes out non-empty first.

Includes a unit-test tweak to confirm this.