QI.bu6oeer: fix creation details after undelete
Previously, when you undeleted a Thing, it would be *mostly* right, but the _creator would be set to the person doing the undelete, and the _createTime would be set to the time of undeletion. This improves the pathways so that we can get it right.

Includes a scenario test. This revealed that we have clearly never had any tests for History, because it was broken: it was hardcoded to the Cassandra stack, so even in the unit-test environment, when we are using the in-memory Akka Persistence implementation, it was still reading the *history* from Cassandra -- which is of course a completely different Space. So I tweaked things to be sensitive to this situation.

Also includes a new `testAssert()` function, to fit in with the TestOp environment. It's probably unnecessary, but makes me more comfortable.