QI.7w4ger8 -- provide a way to undelete arbitrary Things by OID
In and of itself, this isn't enormously useful: it's significant hassle to find the OID, and probably easier to just do it via the History page. But it is a stepping-stone to building a Deleted Things page that provides "undelete" buttons.

Introduces a tweak or two to make it possible to do History management via QL.

Includes a scenario test for the new _undeleteThing() function.