Add fs2
Not sure I'm going to use it right now (I think I'm taking Search in a different direction), but we want it in the stack, so let's bring it in as a dependency now. This version is the last one compatible with Scala 2.11.

Also brings in a decently modern Scalafmt file, so that stuff formats better than the old crap. I'm not going to bother reformatting the world right now, but let's start making it all look nicer.