QI.7w4ger9, QI.7w4gerb -- add _listDeletedThings()
This is a new function that returns a list of all the Things in the Space that have been deleted, optionally including a predicate to use as a filter.

The tickets aren't complete yet -- it needs to be made more powerful (including the OID and deletion time); we need a default page for this; and it needs tests. But the core functionality is now working.

Also includes a bunch of reformatting, since I'm getting serious about scalafmt. There's going to be a lot of that to come.