QI.7w4ger9, QI.7w4gerb -- refactor QLProcessor out of QLParser
This is kind of a massive side-step, but I think push has finally come to shove for this refactor that I have been putting off for years. The motivation is that I want to stop having to recreate a QLText just to pass it through to the SpaceHistory, only to re-parse it -- that's stupidly inefficient. But it was necessary because QLParser is fundamentally about the QLText.

So this rips the "processor" part out of QLParser into its own class. It's still not as FP-ish as I'd like, but it at least helps break down the horrible coupling a *little* bit.

Also, lots more reformatting.