QI.7w4ger9, QI.7w4gerb -- push rendering down into SpaceHistory
Added a new `render` parameter to `_listDeletedThings`, to allow the caller to say exactly what they want to get out of this. That removes the arbitrary display code, and instead puts it in the hands of the user, where it belongs.

Changed the name back to `_listDeletedThings`, since the "rendering" now allows you to return arbitrary types, not just display.

Reworked `_undeleteThing` so that the parameter is optional; it can receive the OIDs instead. This makes more sense in the context of trying to render it.

Fixed ParamResult to think in terms of QLProcessor rather than QLParser.

Fixed some knock-on bugs in InvocationImpl that show up if you are coming in via `processExp()`, with no QLParser.

Still to do:

* Create the standard Deleted Things page
* Tests!