QI.bu6oguy: let's try upgrading Fastparse
The bug at hand is that importing the SI Playtest Space brings Querki to its knees; on the live site, it usually fails.

The changes to ImportSpaceActor are to be a *little* better-managed around errors -- we at least catch and log failures, which didn't happen in many cases before.

That revealed that we are failing in the *parser* for the imported file. Worse, we are failing non-deterministically, which is horrifying. It isn't at all clear what's going on.

That said, we've using using a completely antique version of Fastparse. So let's upgrade to Fastparse2, which will hopefully bring better performance and reliability. This is a big change, because the API changed dramatically.

After this, we should take a look at introducing a lot of cuts, which will hopefully improve memory usage and maybe fix this problem.

Also, deleted the ancient jquery-ui custom tree, which AFAIK we never really used. It hasn't been touched in forever, and dependabot is reporting security issues with it, so let's just scrag it.