QI.bu6oguy: actually fix upload
It turns out that the problem had nothing at all to do with the parser. Instead, to my astonishment, it looks like `Vector[Byte].toArray` is **non-deterministic**! It was producing Arrays of wildly varying length, frequently much shorter than the Vector. Which is bizarre and scary, but easy to fix.

So replaced that `Vector` with an `ArrayBuilder` instead: it's much more efficient *and* more reliable. It's totally mutable and anti-FP, but in our Akka state machine that's the least of our concerns.