QI.bu6oguy: deal with spurious timeout errors
The previous commit got to the point where uploads were working, but we were getting spurious errors along the way, because the frontend's attempts to fetch the state were splashing against the Actor, which was spending all its time parsing the Space. (Which in the case of the Playtest Space, takes a good 20 seconds.) So I moved the parse task into a Future, looping back when done -- that way, it still takes a while, but doesn't kick out errors in the process.

Also improved the progress messages. It looks like I built that system halfway, but never properly finished it because the gap was never apparently in little Spaces. It matters for huge ones, though.

The result of this is that the whole process is now async, so we can get rid of the vestigial synchronous try/catch -- all errors are now coming via the async channel.