QI.7w4gew0: Maintain OID mapping during imports, and use it
The goal here is that, when you have an Imported Space, references to the *old* OIDs will still work. We accomplish that by adding an OldOID Property to each Thing as it gets Imported, and checking that as a fallback during SpaceState lookups.

This probably doesn't cover every possible base yet, but it seems to work for at least some standard smoketests.