QI.9v5kg2b: Step one - make findAllDeleted() dynamic
This is the beginning of the process of rewriting SpaceHistory to not store the full history in RAM. (Which has been OOM'ing the system.)

The first step is the easiest, to deal with findAllDeleted(). This was already phrased in terms of a fold operation, so it's easy to rewrite in the needed way.

In part, this is easiest because we already have pretty good tests for it, so we don't need to add any additional testing, just make sure the existing ones aren't broken.

In order to support that (and everything else we're going to be doing), added the main foldOverHistory() abstraction, which is a wrapper around the Akka Persistence runFoldAsync() with the boilerplate wrapped up. The theory is that everything will eventually be written in terms of that.