QI.9v5kg2b: Add scenario test for walking around in history
Before I go screwing around with listing the history summary, viewing history at a point, and rolling it back, we need to finally get some actual testing of that.

The main point of all this is HistoryMidTests.testHistoryWalking -- everything else is in service to that.

That requires finally implementing the ability for the scenario test harness to pass page parameters to the backend, so that we can pass the desired history version around.

In order to support *that*, took the hardcoded "_historyVersion" param and changed it to a proper val in all the places it is needed.