QI.9v5kg2b: Optimize wandering around in history
Since GetVersion is now fully streamed, we don't want to have to do that for every operation if the user is wandering around in a history version. So instead, cache that particular history version in the OldUserSpaceSession while they are using it.

It's not as optimized as I might like -- the front end is sending two parallel requests when you enter the history version, and those wind up kicking off parallel loads, which is kind of unfortunate. But after that it works as intended; hopefully it's good enough.