QI.9v5kg2b: Remove FindAllStomped
Not strictly necessary (this was streamed to begin with), but while we're in this code, let's remove this command. It was for dealing with a bug that was a number of years ago, and I don't believe anybody has used it in at least a few years. So let's just consider this dead.