QI.9v5kg2b: Delete all the dead code
This is the climax of this ticket: deleting the `history` member (which is what I believe was OOM'ing us), and the function that built it.

Along the way, kill a bunch of imports that we no longer need in the Actor itself, and several interfaces that we turned out to just not be using any more.