Diagnostics and possible self-repair during Space initialization
It looks like the SI Playtest Space is frequently (but not always) failing during load. It isn't clear what's going on. So as one step, let's wrap everything in Exception handlers to make sure that we properly log Exceptions, and deliberately self-destruct the Space so that we can try again.

This isn't necessarily going to help, but should at least help narrow down the possibilities.

Once again, this contains a lot of reformats, that unfortunately swamp the actual semantic changes. (Which are fairly limited.)