Redo upload to be a *little* less ridiculous
Eric is still having problems importing his enormous Playtest Space. This seems to be because, no really, ArrayBuffer.toArray erratically craps out when dealing with an enormous buffer.

So fine: let's rewrite this in terms of streams. It isn't the *right* streams, mind, and we're still tending to reify it into memory before doing anything with it, but it's at least *slightly* closer to correct.