QI.9v5kg2v: Actually redirect
This introduces a new `RedirectFilter`, which watches for request from a "from" host and redirects to a "to". In production, we'll use this to redirect everything from `www.querki.net` to `querki.net`.

It's a bit unfortunate to have to introduce this in front of every request, but in practice it shouldn't introduce a significant amount of latency.