QI.7w4ghfr: add the _toSet function
This deals with a use case that has arisen for Eric: he is winding up with a matrix-y use of `_withValueIn()` that is producing *scads* of duplicates. This new function takes an input List and turns it into a Set, neatly eliminating the dupes.

Includes a couple of simple tests.