Update email to TLS 1.2
AWS has been sending me warnings that we're using TLS 1.1 for email, and they are phasing that out. I *think* (fingers crossed) that that is just a straightforward upgrade.

See https://stackoverflow.com/questions/47166425/how-to-force-javamailsenderimpl-to-use-tls1-2 for more info -- that's where I'm getting the upgrade info from. (SO may be evil, but they're still useful.)