Rewrite config
Splits local.conf and prod.conf out of application.conf.

Introduce the notion of QuerkiEnv, to distinguish the local vs prod environments. (Included test as well, on the theory that we should add that.)

Actually check in application.conf for the first time, since it no longer contains any secrets.