Fix Scenario Tests
The scenario tests (the "Mid Tests" in the older parlance) haven't been kept up with the Querki3 evolution. Rework things to make that function.

This involves a new config file specifically for those, factoring out the SecretsLoader, and making it so that we don't try to load secrets during scenario testing.

That done, we no longer need to hard-code the config settings for the scenario tests -- they just go into the new scenario.conf.