Remove use of Play Crypto
This is a long-standing wart: this class is deprecated in Play 2.5, and removed in 2.6.

Leaving it using CookieSigner for now, which isn't really ideal but is adequate for current use cases. (This code has not been used for passwords for many years.) We'll split that off and do something better later.