Upgrade sbt to 0.13.15
This is the minimum version required by Play 2.6, so let's move this along.

Wound up breaking a bunch of Scaladoc comments that included names like `$item`. I suspect this is a transient bug in sbt; for now, just un-Scaladoc'ed those comments.

Also had to switch from the now-deprecated `persistLauncher` (in SJS) to `scalaJSUseManModuleInitializer`.