Switch upload from Iteratees to Accumulator
The Iteratees stuff gets ripped out into an external library with Play 2.6, and then quickly abandoned. And it's not trivially adapted to our code: the new library changes enough that it's not a one-liner.

So it's time to bite the bullet and rewrite that code in Akka Streams, as probably should have happened almost ten years ago.

The resulting code is *much* more compact, because Akka Streams contains a function (`Sink.actorRefWithAck()`) that behaves very similarly to the hand-built system I had constructed in the early days of Querki. So we're able to completely elide the StreamSendActor, and reduce the complexity of the rest.

Also cleaned up some unused imports, while I notice them.