Suppress spurious error on ApiException
The current version of upickle.read() not only throws (as expected) when it can't resolve the desired type, it spews an error to the log. We've been using that pattern -- catching that and parsing it as an ApiException instead -- but that's always been bad style, and the log errors are just annoying noise.

So rewrite this to work in a slightly more modern fashion instead. (Basically, implement a poor man's validate().)