Upgrade upickle and pprint (but mostly upickle)
This is an enormous upgrade in terms of code impact, because upickle got rid of its full-auto derivation macro, going instead to the same semi-auto that the world has gradually standardized on. This means that every single case class sent on the client/server pipe needs its own `RW` instance, and that's a lot.

(Someday we will have Scala 3, and all of this will turn into nice concise `derives` clauses.)