Upgrade to Play 2.6.25
This is the last release of 2.6, which seems like a good place to go before we do the Scala 2.12 jump.

Somewhere along the way, Play started poking RUNNING_PID into a file, which on Docker winds up located at /opt/docker/RUNNING_PID. Problem is, we don't have write access there. Figured out how to change the location of that, and since we don't actually care about it (I don't *think* it's possible to wind up running multiple instances in the Docker environment), just set that to /dev/null.