Upgrade base64
Turns out that we were still including utest transitively through base64, and this upgraded it past a major API change. That broke FutureCrashTest.scala -- which only existed because, way back in the beginning of time, I had to confirm a workaround for working with utest. But that was the only remaining *use* of utest, so we can just remove it entirely.