SJS 1.13.2 and lots of warnings fixed
In the middle of upgrading Scala.js when I realized I wasn't setting my compiler flags reliably. Fixed that, and then discovered a *ton* of feature flag warnings.

Some of the warnings are postfixOps, which mostly reflects the fact that the language has shifted away from dot-free syntax.  For those, I'll just add the dots.  (Most of it is, eg, places where we are saying "10 seconds", and I've already gotten used to adding the dot in other codebases.)

Some are for implicit conversions.  *That* I think I'll just enable across the board by adding the compiler flag – I use them fairly judiciously, and don't want the headache of adding language imports all over the place for them.

Some are for higher-kinded types.  I'm startled that that's a warning, and it looks like the warning went away during the 2.13 series, so I'll just turn on the compiler flag for now.

A bunch are for existential types, and *that* one is a little distressing.  They're basically all for places where we are extracting a Type or Prop with unknown parameters, and that apparently requires the existentials import.  We need that in way too many places; think about whether we can somehow make it better.  Note that existential types were dropped in Scala 3 (which is part of why I'm going to the hassle of explicitly importing it, so I can find the usages); consider whether something like the specified pattern suggested there would work.

Hit one nasty reflectiveCalls warning. That one looks like a genuine issue: it works, but poorly. Marked it as a TODO to find a better approach.

Also discovered a few more `/:` deprecation warnings that had previously been missed because we weren't doing the warnings for the shared code. Refactored my compiler options to be consistent going forward.