Fix photo upload
It looks like the new Akka machinery we're using for streaming inputs isn't working quite the way the old stuff did. (Specifically, it looks like the Accumulator is returning *before* the upload is complete.)

Fixed this crudely at the receiving end, to be able to cope with the race condition; at some point, I need to go back and figure out how to fix it more correctly.

Along the way, altered the debugging mechanism to dump the photo content to the log instead of a file, since files aren't really a thing in the new world.

Also, discovered that IDEA's unhappiness with the PhotoUploadActor was because of a duplicate (and unnecessary) import of querki.ecology.