Akka 2.5.26
Bring Akka up to the correct version for the current Play. Fix a couple of deprecations caused by that.

Also start dealing with the upcoming deprecation of Logger.error/warn/info/etc, coming in Play 2.7. Time to seriously redo our logging approach to be more modern.

Also pushed akka-persistence-cassandra up a few notches. This requires turning off `meta-in-events-by-tag-view` -- the machinery has changed and needs DB evolution, but we don't actually use tags at all. (Indeed, I find tags a weird non-sequitur in a CQRS application.)

(Note that this started as an attempt to move to Play 2.7.0, which crashed and burned at Akka startup. So we're trying to bring things more up to current before we go there.)