Upgrade akka-persistence-cassandra to 0.98
This is one major step towards being able to upgrade the Cassandra clusters.

At boot time, this performs an adequately-idempotent schema evolution to the Cassandra DB -- this seems to work in local and Test, so I think it's *probably* safe in Prod. The only major concern is whether large Spaces will take a while to load the first time; hopefully not, but we'll see.

Also switches the Java 8 base image, since the one we've been using is deprecated.